refactor: moved `flattenColorObject` to `utils`
